refactor: move data to interface-adapters
