+++ src/get-saml-metadata/src/interface-adapters/data/adapters/FileValidatorAdapter.spec.ts
