+++ src/get-saml-metadata/src/interface-adapters/data/adapters/FileXmlMetadataLoaderAdapter.ts
@@ -1,7 +1,7 @@
 import { readFileSync } from 'fs'
+import { XmlMetadata } from '../../../entities/value-objects/XmlMetadata'
+import { IXmlMetadataLoaderRepository } from '../../../use-cases/ports/IXmlMetadataLoaderRepository'
 import { makeXmlMetadata } from '../factories/makeXmlMetadata'
-import { IXmlMetadataLoaderRepository } from '../../use-cases/ports/IXmlMetadataLoaderRepository'
-import { XmlMetadata } from '../../entities/value-objects/XmlMetadata'
 
 /**
  * Creates a File Loader using `fs` that implements IXmlMetadataLoaderRepository
