+++ src/get-saml-metadata/src/interface-adapters/data/adapters/XmlValidatorAdapter.spec.ts
@@ -1,6 +1,6 @@
 import * as parser from 'fast-xml-parser'
-import { XmlMetadata, XmlMetadataProps } from '../../entities/value-objects/XmlMetadata'
-import { IValidator } from '../../use-cases/ports/IValidator'
+import { XmlMetadata, XmlMetadataProps } from '../../../entities/value-objects/XmlMetadata'
+import { IValidator } from '../../../use-cases/ports/IValidator'
 import { XmlValidatorAdapter } from './XmlValidatorAdapter'
 
 const validXmlString = '<root>Any valid xml</root>'
