+++ src/get-saml-metadata/src/interface-adapters/data/factories/makeFileLoaderAdapter.ts
@@ -1,5 +1,5 @@
 import { FileXmlMetadataLoaderAdapter } from '../../data/adapters/FileXmlMetadataLoaderAdapter'
-import { IXmlMetadataLoaderRepository } from '../../domain/utils/IXmlMetadataLoaderRepository'
+import { IXmlMetadataLoaderRepository } from '../../use-cases/ports/IXmlMetadataLoaderRepository'
 
 export const makeFileXmlMetadataLoaderAdapter = (): IXmlMetadataLoaderRepository => {
   return new FileXmlMetadataLoaderAdapter()
