+++ src/get-saml-metadata/src/use-cases/GetExternalDataInteractor.spec.ts
@@ -1,8 +1,8 @@
 import { fakeMetadata, validMetadataString, validXmlMetadata } from '../../../testdata/fakes'
-import { makeXmlMetadata } from '../data/factories/makeXmlMetadata'
 import { IExternalData } from '../entities/IExternalData'
 import { IMetadata } from '../entities/IMetadataTypes'
 import { XmlMetadata } from '../entities/value-objects/XmlMetadata'
+import { makeXmlMetadata } from '../interface-adapters/data/factories/makeXmlMetadata'
 import { GetExternalDataInteractor } from './GetExternalDataInteractor'
 import { GetExternalDataRequestModel } from './GetExternalDataRequestModel'
 import { GetExternalDataResponseModel } from './GetExternalDataResponseModel'
